server/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pymongo==4.6.0
motor==3.3.2
pydantic==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.2.0
python-multipart==0.0.7
python-dotenv==1.1.1
